Deduplicated the selenium methods by adding a new common class for selenium methods
